Changes related to Foursight React; previously react-experiment branch and PR-88 but redid when rebase from master got confused.
