+++ CHANGELOG.rst
@@ -7,6 +7,21 @@ foursight-cgap
 Change Log
 ----------
 
+2.2.0
+=====
+* Changes related to Foursight React.
+  * Renamed chalicelib directory to chalicelib_cgap.
+  * Renamed target package (pyproject.toml) from chalicelib to chalicelib_cgap.
+  * Moved all Chalice routes to foursight-core (same with foursight).
+  * Moved schedules to chalicelib_cgap/check_schedules.py.
+  * Added chalicelib_local with a sample check_setup.json suitable for local testing.
+  * Changed check_setup.json lookup (in chalicelib_cgap/app_utils.py) to look for the
+    above local file if CHALICE_LOCAL environment variable set to "1"; and also to look
+    for check_setup.json in the directory specified by FOURSIGHT_CHECK_SETUP_DIR if set,
+    otherwise look in the local chalicelib_cgap directory; and setup a fallback directory
+    for this lookup to this local chalicelib_cgap directory, which foursight-core will
+    use if there is no (non-empty) check_setup.json in the specified directory.
+
 2.1.4
 =====
 * Bring in mamga v1.1.0
