+++ chalicelib_cgap/app_utils.py
@@ -1,6 +1,6 @@
 import os
-from os.path import dirname
-from foursight_core.app_utils import AppUtilsCore as AppUtils_from_core
+from foursight_core.app_utils import app  # Chalice object
+from foursight_core.app_utils import AppUtils as AppUtils_from_core
 from foursight_core.identity import apply_identity_globally
 from .vars import FOURSIGHT_PREFIX, HOST
 
@@ -24,8 +24,15 @@ class AppUtils(AppUtils_from_core):
     prefix = FOURSIGHT_PREFIX
     FAVICON = 'https://cgap-dbmi.hms.harvard.edu/favicon.ico'
     host = HOST
-    package_name = 'chalicelib'
-    check_setup_dir = dirname(__file__)
-    DEFAULT_ENV = os.environ.get("ENV_NAME", "cgap-uninitialized")
-    #html_main_title = f'Foursight-{DEFAULT_ENV}'.title().replace("Cgap", "CGAP")
+    package_name = 'chalicelib_cgap'
+    check_setup_dir_fallback = os.path.abspath(os.path.dirname(__file__))
+    if os.environ.get("CHALICE_LOCAL") == "1":
+        check_setup_dir = "chalicelib_local"
+    else:
+        check_setup_dir = os.environ.get("FOURSIGHT_CHECK_SETUP_DIR") or os.path.dirname(__file__)
+    check_setup_dir = os.path.abspath(check_setup_dir)
+    DEFAULT_ENV = os.environ.get("ENV_NAME", "foursight-cgap-env-uninitialized")
     html_main_title = "Foursight" # Foursight CGAP vs Fourfront difference now conveyed in the upper left icon.
+
+
+app_utils_obj = AppUtils.singleton(AppUtils)
