+++ chalicelib_cgap/vars.py
