+++ pyproject.toml
@@ -1,18 +1,18 @@
 [tool.poetry]
 name = "foursight-cgap"
-version = "2.1.4"
+version = "2.1.0.1b81"
 description = "Serverless Chalice Application for Monitoring"
 authors = ["4DN-DCIC Team <support@4dnucleome.org>"]
 license = "MIT"
 packages = [
-  { include = "chalicelib" }
+  { include = "chalicelib_cgap" }
 ]
 
 [tool.poetry.dependencies]
 python = ">=3.7,<3.8"
 dcicutils = "^4.0.2"
 click = "^7.1.2"
-PyJWT = "1.5.3"
+PyJWT = "2.5.0"
 Jinja2 = "2.10.1"
 google-api-python-client = "^1.12.5"
 geocoder = "1.38.1"
@@ -23,7 +23,7 @@ pytz = "^2020.1"
 magma-suite = "1.1.0"
 tibanna-ff = "^1.0.0"
 MarkupSafe = "1.1.1"
-foursight-core = "^2.0.0"
+foursight-core = "2.1.0.1b71"
 # use below for tests but not for deployment
 #foursight-core = { git = "https://github.com/4dn-dcic/foursight-core.git", branch="core2" }
 cgap-pipeline-utils = "23.0"
