+++ tests/checks/test_lifecycle_checks.py
@@ -1,4 +1,4 @@
-from chalicelib.checks.helpers.lifecycle_utils import get_file_lifecycle_status, \
+from chalicelib_cgap.checks.helpers.lifecycle_utils import get_file_lifecycle_status, \
     check_file_lifecycle_status, STANDARD, INFREQUENT_ACCESS, GLACIER, DEEP_ARCHIVE, DELETED
 import json, datetime
 from unittest.mock import patch
@@ -35,7 +35,7 @@ class TestLifecycleChecks():
         return datetime.datetime(2022, 5, 24)
 
 
-    @patch('chalicelib.checks.helpers.lifecycle_utils.get_datetime_utcnow')
+    @patch('chalicelib_cgap.checks.helpers.lifecycle_utils.get_datetime_utcnow')
     @patch('dcicutils.ff_utils.get_metadata')
     @patch('dcicutils.ff_utils.search_metadata')
     def test_check_file_lifecycle_status(self, mock_search_metadata, mock_get_metadata, mock_datetime_utcnow):
