+++ tests/test_fs_connection.py
@@ -20,7 +20,7 @@ class TestFSConnection():
         assert (self.connection.ff_keys is None)
 
     def test_run_check_with_bad_connection(self):
-        check_handler = check_utils.CheckHandler(FOURSIGHT_PREFIX, 'chalicelib', os.path.dirname(chalicelib_path))
+        check_handler = check_utils.CheckHandler(FOURSIGHT_PREFIX, 'chalicelib_cgap', os.path.dirname(chalicelib_path))
         check_res = check_handler.run_check_or_action(self.connection, 'wrangler_checks/item_counts_by_type', {})
         # run_check_or_action returns a dict with results
         print("check_res=" + str(check_res))
